Fix include guards and some more namespace and build issues

+++ src/sensesp/system/minimal_button.h
@@ -36,7 +36,7 @@ class MinimalButtonHandler : public BaseButtonHandler {
         break;
       case AceButton::kEventLongReleased:
         ESP_LOGD(__FILENAME__, "Long release, duration: %d",
-                 time_since_press_event);
+                 static_cast<int>(time_since_press_event));
         if (time_since_press_event > 5000) {
           this->handle_very_long_press();
         } else if (time_since_press_event > 1000) {
@@ -59,7 +59,7 @@ class MinimalButtonHandler : public BaseButtonHandler {
    */
   virtual void handle_button_press() {
     ESP_LOGD(__FILENAME__, "Short release, duration: %d",
-             time_since_press_event);
+             static_cast<int>(time_since_press_event));
     ESP_LOGD(__FILENAME__, "Restarting");
     ESP.restart();
   }
